vähän parannuksia

+++ src/services/gamemgmt.py
@@ -1,8 +1,8 @@
 import json
 import os
 import shutil
-from objects import Game
-from dir import src
+from misc.objects import Game
+from misc.dir import src
 
 
 class GameMgr:
@@ -14,13 +14,15 @@ class GameMgr:
         Args:
             title: the title name of whatever game you want to create
         """
-        gameid = len(os.listdir(src("games/")))+1
-        os.mkdir(src(f"games/{gameid}"))
+        if self.check_if_game_exists(title):
+            return None
+        gameid = len(os.listdir(src("data/games/")))+1
+        os.mkdir(src(f"data/games/{gameid}"))
 
-        with open(src(f"games/{gameid}/gameinfo.json"), "w", encoding="utf-8") as info:
+        with open(src(f"data/games/{gameid}/gameinfo.json"), "w", encoding="utf-8") as info:
             json.dump(vars(Game(title, gameid)), info)
 
-        with open(src(f"games/{gameid}/scores.csv"), "w", encoding="utf-8") as something:
+        with open(src(f"data/games/{gameid}/scores.csv"), "w", encoding="utf-8") as something:
             something.close()
 
     def game_json_list(self):
@@ -31,9 +33,9 @@ class GameMgr:
             list: Sorted list of the content in each games.json.
         """
         templist = []
-        listdir = os.listdir(src("games"))
+        listdir = os.listdir(src("data/games"))
         for gameid in listdir:
-            with open(src(f"games/{gameid}/gameinfo.json"), "r", encoding="utf-8") as file:
+            with open(src(f"data/games/{gameid}/gameinfo.json"), "r", encoding="utf-8") as file:
                 templist.append(json.load(file))
         return sorted(templist, key=lambda game: game["id"])
 
@@ -46,14 +48,29 @@ class GameMgr:
         Returns:
             list: Returns the variables in gameinfo.json in a list with only this element in it.
         """
-        with open(src(f"games/{gameid}/gameinfo.json"), "r", encoding="utf-8") as gameinfo:
+        with open(src(f"data/games/{gameid}/gameinfo.json"), "r", encoding="utf-8") as gameinfo:
             return json.load(gameinfo)
 
     def delete_all_games(self):
         """This function is used for setting up the program and also in tests.
         """
         try:
-            shutil.rmtree(src("games"))
+            shutil.rmtree(src("data/games"))
         except:
             pass
-        os.mkdir(src("games"))
+        os.mkdir(src("data/games"))
+
+    def check_if_game_exists(self, title:str):
+        """Checks to see if a game has already been created with the same name
+
+        Args:
+            title (str): Title/name of the game
+
+        Returns:
+            Boolean : True of False depending on if it has been created
+        """
+        gamelist = self.game_json_list()
+        for game in gamelist:
+            if title.lower()==game["title"].lower():
+                return True
+        return False
